requirements.txt
fastapi
uvicorn
celery
pydantic<2.0.0
python-dotenv
openai
numpy
pgvector
langchain
httpx
sqlmodel==0.0.8
psycopg2-binary
sqlalchemy==1.4.41
